+++ lib/obsobject.dart
@@ -1,3 +1,4 @@
+export 'src/cancelablethen.dart';
 export 'src/commission.dart';
 export 'src/computed.dart';
 export 'src/observable.dart';