+++ lib/src/commission.dart
@@ -1,9 +1,10 @@
 import 'computed.dart';
+import 'observablebase.dart';
 
 ///This class inherit from Computed
 ///and provide a function to set property 'value'
 //Simple to use a Computed similar as an Observable
-class Commission<T> extends Computed<T> {
+class Commission<T> extends Computed<T> with ObservableWritable<T> {
   ///Function used to compute value
   final T Function() reader;
 
@@ -16,7 +17,8 @@ class Commission<T> extends Computed<T> {
         assert(writer != null),
         super(reader);
 
-  set value(T val) {
+  @override
+  void set value(T val) {
     if (val != super.value) writer(val);
   }
 }